+++ src/widgets/ruleswidget.hpp
@@ -15,7 +15,7 @@ class RulesWidget : public QWidget
 {
     Q_OBJECT
 
-    enum Toggles{ svrPassword = 0, world = 1, url = 2, allPK = 3, maxP = 4, maxAFK = 5,
+    enum Toggles{ svrPassword = 0, world = 1, url = 2, allPK = 3, maxP = 4, maxIdle = 5,
                   minV = 6, ladder = 7, noBleep = 8, noCheat = 9,
                   noEavesdrop = 10, noMigrate = 11, noMod = 12, noPets = 13,
                   noPK = 14, arenaPK = 15 };
@@ -25,7 +25,7 @@ class RulesWidget : public QWidget
     SelectWorld* selectWorld{ nullptr };
     bool maxPlayersCheckState{ false };
     bool minVersionCheckState{ false };
-    bool maxAFKCheckState{ false };
+    bool maxIdleCheckState{ false };
     bool worldCheckState{ false };
     bool urlCheckState{ false };
     bool pwdCheckState{ false };